requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
langchain>=0.2.12
langchain-google-genai>=0.0.9
langchain-core>=0.2.5
google-genai>=1.0.0
